Create inspect pkg
Some of the structs in p2-inspect may be useful elsewhere (ie p2-show-health),
so I have moved them into their own package.
