+++ bin/p2-inspect/main.go
@@ -2,14 +2,12 @@ package main
 
 import (
 	"encoding/json"
-	"fmt"
 	"log"
 	"net/http"
 	"os"
-	"sort"
-	"strings"
 
 	"github.com/square/p2/pkg/health"
+	"github.com/square/p2/pkg/inspect"
 	"github.com/square/p2/pkg/kp"
 	"github.com/square/p2/pkg/util/net"
 	"github.com/square/p2/pkg/version"
@@ -24,26 +22,6 @@ var (
 	headers        = kingpin.Flag("header", "An HTTP header to add to requests, in KEY=VALUE form. Can be specified multiple times.").StringMap()
 	https          = kingpin.Flag("https", "Use HTTPS").Bool()
 )
-
-const (
-	INTENT_SOURCE = iota
-	REALITY_SOURCE
-)
-
-type NodePodStatus struct {
-	NodeName           string             `json:"node,omitempty"`
-	PodId              string             `json:"pod,omitempty"`
-	IntentManifestSHA  string             `json:"intent_manifest_sha"`
-	RealityManifestSHA string             `json:"reality_manifest_sha"`
-	IntentLocations    []string           `json:"intent_locations"`
-	RealityLocations   []string           `json:"reality_locations"`
-	Health             health.HealthState `json:"health,omitempty"`
-}
-
-type NodeHealth struct {
-	Status string `json:"status"`
-	Output string `json:"output"`
-}
 
 func main() {
 	kingpin.Version(version.VERSION)
@@ -76,16 +54,16 @@ func main() {
 		}
 	}
 
-	statusMap := make(map[string]map[string]NodePodStatus)
+	statusMap := make(map[string]map[string]inspect.NodePodStatus)
 
 	for _, kvp := range intents {
-		if addKVPToMap(kvp, INTENT_SOURCE, *filterNodeName, *filterPodId, statusMap) != nil {
+		if inspect.AddKVPToMap(kvp, inspect.INTENT_SOURCE, *filterNodeName, *filterPodId, statusMap) != nil {
 			log.Fatal(err)
 		}
 	}
 
 	for _, kvp := range realities {
-		if addKVPToMap(kvp, REALITY_SOURCE, *filterNodeName, *filterPodId, statusMap) != nil {
+		if inspect.AddKVPToMap(kvp, inspect.REALITY_SOURCE, *filterNodeName, *filterPodId, statusMap) != nil {
 			log.Fatal(err)
 		}
 	}
@@ -110,51 +88,4 @@ func main() {
 
 	enc := json.NewEncoder(os.Stdout)
 	enc.Encode(statusMap)
-}
-
-func addKVPToMap(result kp.ManifestResult, source int, filterNode, filterPod string, statuses map[string]map[string]NodePodStatus) error {
-	keySegs := strings.Split(result.Path, "/")
-	nodeName := keySegs[1]
-	podId := keySegs[2]
-
-	if filterNode != "" && nodeName != filterNode {
-		return nil
-	}
-	if filterPod != "" && podId != filterPod {
-		return nil
-	}
-
-	if statuses[podId] == nil {
-		statuses[podId] = make(map[string]NodePodStatus)
-	}
-	old := statuses[podId][nodeName]
-
-	manifestSHA, err := result.Manifest.SHA()
-	if err != nil {
-		return err
-	}
-
-	switch source {
-	case INTENT_SOURCE:
-		if old.IntentManifestSHA != "" {
-			return fmt.Errorf("Two intent manifests for node %s pod %s", nodeName, podId)
-		}
-		old.IntentManifestSHA = manifestSHA
-		for _, launchable := range result.Manifest.LaunchableStanzas {
-			old.IntentLocations = append(old.IntentLocations, launchable.Location)
-		}
-		sort.Strings(old.IntentLocations)
-	case REALITY_SOURCE:
-		if old.RealityManifestSHA != "" {
-			return fmt.Errorf("Two reality manifests for node %s pod %s", nodeName, podId)
-		}
-		old.RealityManifestSHA = manifestSHA
-		for _, launchable := range result.Manifest.LaunchableStanzas {
-			old.RealityLocations = append(old.RealityLocations, launchable.Location)
-		}
-		sort.Strings(old.RealityLocations)
-	}
-
-	statuses[podId][nodeName] = old
-	return nil
 }
